+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added support for errorbars.
+- Added support for errorbars and line markers.
 
 ## [0.3.0] - 2024-04-30